renaming of variables/ files
